feat: add `BlockingObjectSource`
